Add `yii\base\ComponentTrait`.
