+++ framework/base/ComponentInterface.php
@@ -1,0 +1,160 @@
+<?php
+
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\base;
+
+/**
+ *
+ * @author Robert Korulczyk <robert@korulczyk.pl>
+ * @since 2.0.11
+ */
+interface ComponentInterface extends MutatorsInterface
+{
+    /**
+     * Returns a list of behaviors that this component should behave as.
+     *
+     * Child classes may override this method to specify the behaviors they want to behave as.
+     *
+     * The return value of this method should be an array of behavior objects or configurations
+     * indexed by behavior names. A behavior configuration can be either a string specifying
+     * the behavior class or an array of the following structure:
+     *
+     * ```php
+     * 'behaviorName' => [
+     *     'class' => 'BehaviorClass',
+     *     'property1' => 'value1',
+     *     'property2' => 'value2',
+     * ]
+     * ```
+     *
+     * Note that a behavior class must extend from [[Behavior]]. Behaviors can be attached using a name or anonymously.
+     * When a name is used as the array key, using this name, the behavior can later be retrieved using [[getBehavior()]]
+     * or be detached using [[detachBehavior()]]. Anonymous behaviors can not be retrieved or detached.
+     *
+     * Behaviors declared in this method will be attached to the component automatically (on demand).
+     *
+     * @return array the behavior configurations.
+     */
+    public function behaviors();
+
+    /**
+     * Returns a value indicating whether there is any handler attached to the named event.
+     * @param string $name the event name
+     * @return boolean whether there is any handler attached to the event.
+     */
+    public function hasEventHandlers($name);
+
+    /**
+     * Attaches an event handler to an event.
+     *
+     * The event handler must be a valid PHP callback. The following are
+     * some examples:
+     *
+     * ```
+     * function ($event) { ... }         // anonymous function
+     * [$object, 'handleClick']          // $object->handleClick()
+     * ['Page', 'handleClick']           // Page::handleClick()
+     * 'handleClick'                     // global function handleClick()
+     * ```
+     *
+     * The event handler must be defined with the following signature,
+     *
+     * ```
+     * function ($event)
+     * ```
+     *
+     * where `$event` is an [[Event]] object which includes parameters associated with the event.
+     *
+     * @param string $name the event name
+     * @param callable $handler the event handler
+     * @param mixed $data the data to be passed to the event handler when the event is triggered.
+     * When the event handler is invoked, this data can be accessed via [[Event::data]].
+     * @param boolean $append whether to append new event handler to the end of the existing
+     * handler list. If false, the new handler will be inserted at the beginning of the existing
+     * handler list.
+     * @see off()
+     */
+    public function on($name, $handler, $data = null, $append = true);
+
+    /**
+     * Detaches an existing event handler from this component.
+     * This method is the opposite of [[on()]].
+     * @param string $name event name
+     * @param callable $handler the event handler to be removed.
+     * If it is null, all handlers attached to the named event will be removed.
+     * @return boolean if a handler is found and detached
+     * @see on()
+     */
+    public function off($name, $handler = null);
+
+    /**
+     * Triggers an event.
+     * This method represents the happening of an event. It invokes
+     * all attached handlers for the event including class-level handlers.
+     * @param string $name the event name
+     * @param Event $event the event parameter. If not set, a default [[Event]] object will be created.
+     */
+    public function trigger($name, Event $event = null);
+
+    /**
+     * Returns the named behavior object.
+     * @param string $name the behavior name
+     * @return null|Behavior the behavior object, or null if the behavior does not exist
+     */
+    public function getBehavior($name);
+
+    /**
+     * Returns all behaviors attached to this component.
+     * @return Behavior[] list of behaviors attached to this component
+     */
+    public function getBehaviors();
+
+    /**
+     * Attaches a behavior to this component.
+     * This method will create the behavior object based on the given
+     * configuration. After that, the behavior object will be attached to
+     * this component by calling the [[Behavior::attach()]] method.
+     * @param string $name the name of the behavior.
+     * @param string|array|Behavior $behavior the behavior configuration. This can be one of the following:
+     *
+     *  - a [[Behavior]] object
+     *  - a string specifying the behavior class
+     *  - an object configuration array that will be passed to [[Yii::createObject()]] to create the behavior object.
+     *
+     * @return Behavior the behavior object
+     * @see detachBehavior()
+     */
+    public function attachBehavior($name, $behavior);
+
+    /**
+     * Attaches a list of behaviors to the component.
+     * Each behavior is indexed by its name and should be a [[Behavior]] object,
+     * a string specifying the behavior class, or an configuration array for creating the behavior.
+     * @param array $behaviors list of behaviors to be attached to the component
+     * @see attachBehavior()
+     */
+    public function attachBehaviors($behaviors);
+
+    /**
+     * Detaches a behavior from the component.
+     * The behavior's [[Behavior::detach()]] method will be invoked.
+     * @param string $name the behavior's name.
+     * @return null|Behavior the detached behavior. Null if the behavior does not exist.
+     */
+    public function detachBehavior($name);
+
+    /**
+     * Detaches all behaviors from the component.
+     */
+    public function detachBehaviors();
+
+    /**
+     * Makes sure that the behaviors declared in [[behaviors()]] are attached to this component.
+     */
+    public function ensureBehaviors();
+}
