+++ framework/base/Object.php
@@ -7,8 +7,6 @@
 
 namespace yii\base;
 
-use Yii;
-
 /**
  * Object is the base class that implements the *property* feature.
  *
